Project_1_RAG_Knowledge_Assistant/requirements.txt
streamlit
#python-dotenv
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
langchain-text-splitters
chromadb
faiss-cpu
pypdf
docx2txt
python-pptx
openpyxl
unstructured
msoffcrypto-tool
beautifulsoup4
pinecone
langchain-pinecone
langchain-astradb
langchain-weaviate
weaviate-client

